+++ HISTORY.md
@@ -8,6 +8,7 @@ unreleased
   * Send complete HTML document in redirect & error responses
   * Set default CSP header in redirect & error responses
   * Use `res.getHeaderNames()` when available
+  * Use `res.headersSent` when available
   * deps: debug@2.6.1
     - Allow colors in workers
     - Deprecated `DEBUG_FD` environment variable set to `3` or higher